Extract signup_eligibility module from services/praxis.py (#2870)
Cut 3 of #1391: the signup-gate cluster (meets_task_level, evaluate_signup,
gather_signup_facts, signup_reason, allowed_praxis_modes,
is_task_eligible_for_character, the Double Dipper membership machinery, and
the renamed-public count_in_progress_praxes) moves to its own module. Pure
move, no behaviour change — services/praxis.py imports it back for the
pieces its own create/invite flow still needs.
